support current_app as an argument for reverse
current_app has been a argument for reverse since at least django 1.7 and is part of the django template url tag, and is a requirement if you rely on current_app for your url resolving with multiple apps for instance multiple django admins. This code is pretty much exactly how django's template tag does it.